# torneo-futnorte-api/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Torneo FutNorte API
  description: |
    API REST para la gestión completa de torneos de fútbol siguiendo arquitectura hexagonal.
    
    Esta API permite administrar:
    - **Torneos**: Crear, gestionar y controlar el ciclo de vida de torneos
    - **Equipos**: Registrar equipos participantes en torneos
    - **Jugadores**: Gestionar jugadores asociados a equipos
    - **Enfrentamientos**: Programar y registrar partidos entre equipos
    - **Estadísticas**: Seguimiento de goles y estadísticas de jugadores
    
    ## Arquitectura
    Implementa **Arquitectura Hexagonal (Puertos y Adaptadores)** con:
    - Dominio puro sin dependencias externas
    - Casos de uso bien definidos
    - Adaptadores para persistencia y web
    - Sistema de excepciones personalizado con trazabilidad
    
    ## Manejo de Errores
    Todos los errores incluyen:
    - Código de error específico
    - Mensaje descriptivo
    - Trace ID para trazabilidad
    - Timestamp del error
    - Detalles adicionales cuando aplique
  version: 1.0.0
  contact:
    name: Equipo FutNorte
    email: dev@futnorte.com
    url: https://github.com/jotalvaro7/torneo-futnorte-api
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Servidor de Desarrollo
  - url: https://api.futnorte.com/api
    description: Servidor de Producción
paths:
...
  /enfrentamientos:
    post:
      tags:
        - Enfrentamientos
      summary: Crear un nuevo enfrentamiento
      description: Programa un nuevo enfrentamiento entre dos equipos de un torneo
      operationId: crearEnfrentamiento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrearEnfrentamientoRequest'
      responses:
        '201':
          description: Enfrentamiento creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnfrentamientoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /enfrentamientos/{id}:
    get:
      tags:
        - Enfrentamientos
      summary: Obtener enfrentamiento por ID
      description: Recupera la información de un enfrentamiento específico
      operationId: obtenerEnfrentamiento
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del enfrentamiento
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Enfrentamiento encontrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnfrentamientoResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags:
        - Enfrentamientos
      summary: Actualizar enfrentamiento
      description: |
        Actualiza cualquier aspecto de un enfrentamiento incluyendo fecha, hora, cancha,
        estado y resultados. Este endpoint consolidado reemplaza múltiples operaciones separadas.

        **Funcionalidades incluidas:**
        - Cambiar fecha, hora y cancha
        - Finalizar enfrentamiento con goles del equipo
        - Registrar goles individuales por jugador
        - Aplazar enfrentamiento
        - Reprogramar enfrentamiento aplazado
        - Corregir resultados y goles de enfrentamientos finalizados
      operationId: actualizarEnfrentamiento
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del enfrentamiento
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActualizarEnfrentamientoRequest'
            example:
              estado: "FINALIZADO"
              golesLocal: 2
              golesVisitante: 1
              golesJugadoresLocal:
                - jugadorId: 2
                  cantidadGoles: 2
              golesJugadoresVisitante:
                - jugadorId: 12
                  cantidadGoles: 1
      responses:
        '200':
          description: Enfrentamiento actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnfrentamientoResponse'
              example:
                id: 1
                torneoId: 1
                equipoLocal: "Real Madrid"
                equipoVisitante: "Barcelona"
                fechaHora: "2024-03-15T20:00:00"
                cancha: "Estadio Santiago Bernabéu"
                estado: "FINALIZADO"
                golesLocal: 2
                golesVisitante: 1
                golesJugadoresLocal:
                  - jugadorId: 2
                    nombreJugador: "Karim"
                    apellidoJugador: "Benzema"
                    cantidadGoles: 2
                golesJugadoresVisitante:
                  - jugadorId: 12
                    nombreJugador: "Robert"
                    apellidoJugador: "Lewandowski"
                    cantidadGoles: 1
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
        - Enfrentamientos
      summary: Eliminar enfrentamiento
      description: |
        Elimina un enfrentamiento del sistema. Con permisos de administrador,
        se puede eliminar cualquier enfrentamiento sin restricciones de estado.
      operationId: eliminarEnfrentamiento
      parameters:
        - name: id
          in: path
          required: true
          description: Identificador único del enfrentamiento
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '204':
          description: Enfrentamiento eliminado exitosamente
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /enfrentamientos/torneo/{torneoId}:
    get:
      tags:
        - Enfrentamientos
      summary: Obtener enfrentamientos por torneo
      description: Recupera todos los enfrentamientos de un torneo específico
      operationId: obtenerEnfrentamientosPorTorneo
      parameters:
        - name: torneoId
          in: path
          required: true
          description: Identificador único del torneo
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Lista de enfrentamientos del torneo obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnfrentamientoResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /enfrentamientos/equipo/{equipoId}:
    get:
      tags:
        - Enfrentamientos
      summary: Obtener enfrentamientos por equipo
      description: Recupera todos los enfrentamientos donde participa un equipo específico
      operationId: obtenerEnfrentamientosPorEquipo
      parameters:
        - name: equipoId
          in: path
          required: true
          description: Identificador único del equipo
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        '200':
          description: Lista de enfrentamientos del equipo obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnfrentamientoResponse'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /enfrentamientos/fecha:
    get:
      tags:
        - Enfrentamientos
      summary: Obtener enfrentamientos por rango de fechas
      description: Recupera enfrentamientos programados en un rango de fechas específico
      operationId: obtenerEnfrentamientosPorFecha
      parameters:
        - name: fechaInicio
          in: query
          required: true
          description: Fecha y hora de inicio del rango
          schema:
            type: string
            format: date-time
            example: "2024-03-01T00:00:00"
        - name: fechaFin
          in: query
          required: true
          description: Fecha y hora de fin del rango
          schema:
            type: string
            format: date-time
            example: "2024-03-31T23:59:59"
      responses:
        '200':
          description: Lista de enfrentamientos en el rango de fechas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EnfrentamientoResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: Respuesta estándar para errores de la API con trazabilidad
      properties:
        code:
          type: string
          description: Código específico del error para identificación programática
          example: "ENTITY_NOT_FOUND"
          enum:
            - ENTITY_NOT_FOUND
            - DUPLICATE_ENTITY
            - VALIDATION_ERROR
            - BUSINESS_RULE_VIOLATION
            - DATABASE_ERROR
            - ACCESS_DENIED
            - INTERNAL_SERVER_ERROR
        message:
          type: string
          description: Mensaje descriptivo del error para humanos
          example: "Torneo con ID 1 no encontrado"
        traceId:
          type: string
          description: Identificador único para rastrear el error en logs
          example: "a1b2c3d4"
          pattern: "^[a-f0-9]{8}$"
        path:
          type: string
          description: Ruta del endpoint que generó el error
          example: "/api/torneos/1"
        timestamp:
          type: string
          format: date-time
          description: Fecha y hora exacta cuando ocurrió el error
          example: "2024-03-15T10:30:45"
        details:
          type: object
          description: Detalles adicionales del error (opcional)
          additionalProperties:
            type: string
          example:
            field: "nombre"
            reason: "es obligatorio"
      required:
        - code
        - message
        - timestamp
...
    EnfrentamientoResponse:
      type: object
      description: Respuesta detallada de un enfrentamiento con nombres de equipos y goleadores
      properties:
        id:
          type: integer
          format: int64
          description: Identificador único del enfrentamiento
          example: 1
        torneoId:
          type: integer
          format: int64
          description: ID del torneo
          example: 1
        equipoLocal:
          type: string
          description: Nombre del equipo local
          example: "Real Madrid"
        equipoVisitante:
          type: string
          description: Nombre del equipo visitante
          example: "Barcelona"
        fechaHora:
          type: string
          format: date-time
          description: Fecha y hora del enfrentamiento
          example: "2024-03-15T20:00:00"
        cancha:
          type: string
          description: Cancha donde se juega
          example: "Estadio Santiago Bernabéu"
        estado:
          $ref: '#/components/schemas/EstadoEnfrentamiento'
        golesLocal:
          type: integer
          description: Goles del equipo local
          example: 2
          nullable: true
        golesVisitante:
          type: integer
          description: Goles del equipo visitante
          example: 1
          nullable: true
        golesJugadoresLocal:
          type: array
          description: Lista de goleadores del equipo local (solo en enfrentamientos finalizados)
          items:
            $ref: '#/components/schemas/GolesJugadorResponse'
          example:
            - jugadorId: 2
              nombreJugador: "Karim"
              apellidoJugador: "Benzema"
              cantidadGoles: 2
        golesJugadoresVisitante:
          type: array
          description: Lista de goleadores del equipo visitante (solo en enfrentamientos finalizados)
          items:
            $ref: '#/components/schemas/GolesJugadorResponse'
          example:
            - jugadorId: 12
              nombreJugador: "Robert"
              apellidoJugador: "Lewandowski"
              cantidadGoles: 1
      required:
        - id
        - torneoId
        - equipoLocal
        - equipoVisitante
        - fechaHora
        - cancha
        - estado
        - golesJugadoresLocal
        - golesJugadoresVisitante
    EstadoEnfrentamiento:
      type: string
      description: Estados posibles de un enfrentamiento
      enum:
        - PROGRAMADO
        - FINALIZADO
        - APLAZADO
      example: "PROGRAMADO"
    CrearEnfrentamientoRequest:
      type: object
      description: Datos requeridos para crear un nuevo enfrentamiento
      properties:
        torneoId:
          type: integer
          format: int64
          description: ID del torneo
          example: 1
        equipoLocalId:
          type: integer
          format: int64
          description: ID del equipo local
          example: 1
        equipoVisitanteId:
          type: integer
          format: int64
          description: ID del equipo visitante
          example: 2
        fechaHora:
          type: string
          format: date-time
          description: Fecha y hora programada
          example: "2024-03-15T20:00:00"
        cancha:
          type: string
          description: Cancha donde se jugará
          example: "Estadio Santiago Bernabéu"
          maxLength: 100
      required:
        - torneoId
        - equipoLocalId
        - equipoVisitanteId
        - fechaHora
        - cancha
    ActualizarEnfrentamientoRequest:
      type: object
      description: |
        Datos para actualizar un enfrentamiento existente. Todos los campos son opcionales,
        permitiendo actualizaciones parciales y flexibles.
      properties:
        fechaHora:
          type: string
          format: date-time
          description: Nueva fecha y hora del enfrentamiento
          example: "2024-03-16T21:00:00"
        cancha:
          type: string
          description: Nueva cancha donde se jugará
          example: "Estadio Camp Nou"
          maxLength: 100
        estado:
          $ref: '#/components/schemas/EstadoEnfrentamiento'
          description: Nuevo estado del enfrentamiento
        golesLocal:
          type: integer
          description: |
            Goles anotados por el equipo local. Requerido si se cambia estado a FINALIZADO.
          example: 2
          minimum: 0
        golesVisitante:
          type: integer
          description: |
            Goles anotados por el equipo visitante. Requerido si se cambia estado a FINALIZADO.
          example: 1
          minimum: 0
        golesJugadoresLocal:
          type: array
          description: |
            Lista de goles anotados por jugadores del equipo local. Permite registrar
            goleadores específicos junto con el resultado del partido.
          items:
            $ref: '#/components/schemas/GolesJugadorDto'
        golesJugadoresVisitante:
          type: array
          description: |
            Lista de goles anotados por jugadores del equipo visitante. Permite registrar
            goleadores específicos junto con el resultado del partido.
          items:
            $ref: '#/components/schemas/GolesJugadorDto'
    GolesJugadorDto:
      type: object
      description: |
        Representa los goles anotados por un jugador específico en un enfrentamiento.
        Permite un registro detallado de goleadores individuales.
      properties:
        jugadorId:
          type: integer
          format: int64
          description: Identificador único del jugador que anotó los goles
          example: 10
        cantidadGoles:
          type: integer
          description: Cantidad de goles anotados por el jugador en el enfrentamiento
          example: 2
          minimum: 1
      required:
        - jugadorId
        - cantidadGoles
    GolesJugadorResponse:
      type: object
      description: |
        Información detallada de un goleador en un enfrentamiento específico.
        Incluye datos del jugador y cantidad de goles anotados.
      properties:
        jugadorId:
          type: integer
          format: int64
          description: Identificador único del jugador
          example: 2
        nombreJugador:
          type: string
          description: Nombre del jugador goleador
          example: "Karim"
          maxLength: 50
        apellidoJugador:
          type: string
          description: Apellido del jugador goleador
          example: "Benzema"
          maxLength: 50
        cantidadGoles:
          type: integer
          description: Cantidad de goles anotados por el jugador en este enfrentamiento
          example: 2
          minimum: 1
      required:
        - jugadorId
        - nombreJugador
        - apellidoJugador
        - cantidadGoles
  responses:
    BadRequest:
      description: Solicitud inválida - Error en datos de entrada o violación de reglas de negocio
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            ValidationError:
              summary: Error de validación
              value:
                code: "VALIDATION_ERROR"
                message: "El nombre del torneo es obligatorio"
                traceId: "a1b2c3d4"
                path: "/api/torneos"
                timestamp: "2024-03-15T10:30:45"
            BusinessRuleError:
              summary: Violación de regla de negocio
              value:
                code: "BUSINESS_RULE_VIOLATION"
                message: "El torneo solo puede iniciarse si está en estado CREADO O CANCELADO"
                traceId: "b2c3d4e5"
                path: "/api/torneos/1/iniciar"
                timestamp: "2024-03-15T10:35:20"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "ENTITY_NOT_FOUND"
            message: "Torneo con ID 999 no encontrado"
            traceId: "c3d4e5f6"
            path: "/api/torneos/999"
            timestamp: "2024-03-15T10:32:15"
...
    InternalServerError:
      description: Error interno del servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "INTERNAL_SERVER_ERROR"
            message: "Ha ocurrido un error interno en el servidor"
            traceId: "e5f6a7b8"
            path: "/api/torneos"
            timestamp: "2024-03-15T10:34:30"
...